watchers/requirements.txt
# Core
watchdog>=4.0.0
python-dotenv>=1.0.0

# Gmail Watcher
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# WhatsApp & LinkedIn Watchers
playwright>=1.43.0
